pyproject.toml
[project]
name = "token-counter"
version = "0.1.0"
description = "Unified token counter for OpenAI, Anthropic, and Gemini models"
requires-python = ">=3.11.2"
dependencies = [
    "tiktoken>=0.8.0",
    "anthropic>=0.43.0",
    "google-genai>=1.0.0",
    "genai-prices>=0.0.34",
]

[project.scripts]
token-counter = "cli:main"

[project.optional-dependencies]
dev = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
